Fixed the minimum weight bounds in the mimial weight constraints.
I was comparing weights of products of operators to the weights
of the individual factors, when I should have been comparing them
to the minimum weight of any factor  from the original operators
in the product.

I also wrote some tests that check to make sure that the correct
number of solutions is generated --- no more, no less.